Separated library submenus, improved scrolling and selecting song UX
