+++ src/games/cclcc/librarysubmenus.h
@@ -9,23 +9,6 @@
 namespace Impacto {
 namespace UI {
 namespace CCLCC {
-class MusicTrackButton : public Widgets::Button {
- public:
-  MusicTrackButton(int id, int position, glm::vec2 pos);
-
-  void Show() override;
-  void Update(float dt) override;
-  void Render() override;
-  void Move(glm::vec2 relativePosition) override;
-
-  static auto GetSc3LockedText(glm::vec2 pos) {}
-
- private:
-  std::array<ProcessedTextGlyph, 6> LockedText;
-  std::array<ProcessedTextGlyph, 4> NumberText;
-  Widgets::Label ArtistName;
-  int Position = 0;
-};
 class LibrarySubmenu : public Menu {
  public:
   LibrarySubmenu();
@@ -33,6 +16,7 @@ class LibrarySubmenu : public Menu {
   void Show() override;
   void Hide() override;
   void Update(float dt) override;
+  virtual void UpdateInput(float) { Menu::UpdateInput(); };
   void Render() override;
   // void Move(glm::vec2 offset);
   // void MoveTo(glm::vec2 pos);
@@ -45,24 +29,6 @@ class LibrarySubmenu : public Menu {
 class AlbumMenu : public LibrarySubmenu {
  public:
   AlbumMenu() = default;
-};
-
-class MusicMenu : public LibrarySubmenu {
- public:
-  enum class MusicPlayMode { Normal, Shuffle, Repeat, RepeatOne };
-  MusicMenu();
-  void Update(float dt) override;
-  void Render() override;
-  void Show() override;
-
-  float PageY = 0;
-  MusicPlayMode PlayMode = MusicPlayMode::Normal;
-};
-
-class MovieMenu : public LibrarySubmenu {
- public:
-  MovieMenu();
-  void Show() override;
 };
 
 }  // namespace CCLCC
