+++ src/games/cclcc/moviemenu.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "librarysubmenus.h"
+
+namespace Impacto::UI::CCLCC {
+
+class MovieMenu : public LibrarySubmenu {
+ public:
+  MovieMenu();
+  void Show() override;
+};
+
+}
