+++ src/profile/games/cclcc/librarymenu.h
@@ -35,6 +35,7 @@ inline Sprite MusicItemsOverlaySprite;
 inline glm::vec2 MusicItemsOverlayPosition;
 inline std::vector<uint8_t> MusicPlayIds;
 inline RectF MusicRenderingBounds;
+inline RectF MusicHoverBounds;
 inline uint8_t MusicStringTableId;
 inline int MusicStringLockedIndex;
 inline int MusicTrackNameSize;
@@ -42,7 +43,11 @@ inline int MusicTrackNameOffsetX;
 inline int MusicTrackArtistSize;
 inline int MusicTrackArtistOffsetX;
 inline int MusicButtonTextYOffset;
+inline float MusicDirectionalHoldTime;
+inline float MusicDirectionalFocusTimeInterval;
 inline RectF MusicButtonBounds;
+inline Sprite MusicButtonHoverSprite;
+inline Sprite MusicButtonSelectSprite;
 
 inline std::vector<Sprite> MovieDiskSprites;
 inline std::vector<Sprite> MovieDiskHighlightSprites;
